fix: address review feedback on #1929
Trim two doc comments to plain behavior summaries, move the
`derive_reader_display` test out of mod.rs into a sibling tests.rs, add a
data-testid for the header chapter span, and assert the TOC-jump's
progress-save POST in the new Playwright spec instead of a bare class locator.
